scar-judging-django/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# HTTP requests (for Challonge API and Discord webhooks)
requests>=2.31,<3.0

# HTML parsing (for RCE scraping)
beautifulsoup4>=4.12,<5.0

# Production server
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0

# Environment variables
python-dotenv>=1.0,<2.0
